[N1QL] HTTP: Provide clearer semantics for path handling
I ran into this problem because the library was adding a '/' at the end
of the path. This isn't necessarily a good thing and may end up
confusing some servers (specifically, n1ql distinguishes -- in one
incarnation) between /query and /query/. While this may be a bug inside
the host, it was still rather annoying to deal with.

This commit resolves this by not making _path_ required, and only
adds a '/' to path if the prefix does not end with '/' and the path does
not begin with a '/'.

Change-Id: I5a85d7384d4dfe537b4b5898bac1fc8ab331583d