CCBC-566: DNS SRV support
This adds a `dnssrv` option to the connection string. If set, the input
hostname will be selected and attempted for a DNS SRV query. If it
succeeds, then the hosts use for the query will be returned; otherwise
the hostlist remains in tact.

Currently support depends on either Windows (which works "out of the
box") and libbind/libresolv, which come bundled with glibc on both OS X
and Linux. This may fail to function on very ancient systems; in which
event these routines are stubbed out with no-op equivalents.

Change-Id: I0d54b63c1e6cef64e5a09280343e9aba24ed7a80